commit 734345536 in 2019

+++ buildAPKs/init/mod.bash
@@ -7,25 +7,22 @@ shopt -s nullglob globstar
 . "$RDR/scripts/bash/shlibs/trap.bash" 207 208 209 "${0##*/}" "wake.start" 
 
 _CLINKS_() { 
-	ADLINK=(apps buildAPKs buildAPKs.modules clocks compasses developers.tools entertainment flashlights games github live.wallpapers samples top10 tutorials widgets)
+	ADLINK=(apps buildAPKs buildAPKs.modules clocks compasses developers.tools entertainment flashlights games live.wallpapers samples top10 tutorials widgets)
 	VSTRING="symlink warning: Continuing..."
 	VSTRINGC="Creating symlinks: "
 	if [[ ! -e "$RDR/update.buildAPKs.sh" ]]
 	then
 		printf "\\e[1;34m%s" "$VSTRINGC"
-		ln -s "$RDR/scripts/sh/shlibs/buildAPKs/maintenance/up.sh" "$RDR/update.buildAPKs.sh" 2>/dev/null || printf "%s\\n" "update.buildAPKs.sh $VSTRING"
-		ln -s "$RDR/scripts/sh/shlibs/dctar.sh" "$RDR/scripts/maintenance/delete.corrupt.tars.sh" 2>/dev/null || printf "%s\\n" "delete.corrupt.tars.sh $VSTRING"
+		ln -s "$RDR/scripts/sh/shlibs/buildAPKs/maintenance/up.sh" "$RDR/update.buildAPKs.sh" || printf "%s\\n" "update.buildAPKs.sh $VSTRING"
+		[ ! -e  "$RDR/scripts/maintenance/delete.corrupt.tars.sh" ] && ln -s "$RDR/scripts/sh/shlibs/dctar.sh" "$RDR/scripts/maintenance/delete.corrupt.tars.sh" || printf "%s\\n" "delete.corrupt.tars.sh $VSTRING"
 		for TYPE in "${ADLINK[@]}"
 		do
-			ln -s "$RDR/scripts/bash/build/build.$TYPE.bash" "$RDR/build.$TYPE.bash" 2>/dev/null  || printf "%s\\n" "build.$TYPE.bash $VSTRING"
+			[ ! -e "$RDR/build.$TYPE.bash" ] && ln -s "$RDR/scripts/bash/build/build.$TYPE.bash" "$RDR/build.$TYPE.bash" || printf "%s\\n" "build.$TYPE.bash $VSTRING"
 		done
-		ln -s "$RDR/scripts/bash/build/buildAll.bash" "$RDR/buildAll.bash" 2>/dev/null || printf "%s\\n" "buildAll.bash $VSTRING"
-		ln -s "$RDR/var/cache/stash" "$RDR/stash" 2>/dev/null || printf "%s\\n" "stash $VSTRING"
+		[ ! -e "$RDR/build.github.bash" ] && ln -s "$RDR/scripts/bash/github/build.github.bash" "$RDR/build.github.bash" || printf "%s\\n" "build.github.bash $VSTRING"
+		[ ! -e "$RDR/stash" ] && ln -s "$RDR/var/cache/stash" "$RDR/stash" || printf "%s\\n" "stash $VSTRING"
 		printf "\\e[1;32mDONE\\e[0m\\n"
-		if [[ -e "$RDR/setup.buildAPKs.bash" ]]
-		then
-			rm -f "$RDR/setup.buildAPKs.bash"
-		fi
+		[ -e "$RDR/setup.buildAPKs.bash" ] && rm -f "$RDR/setup.buildAPKs.bash"
 	fi
 }
